+++ src/Audit.EntityFramework/DbContextHelper.Net45.cs
@@ -250,6 +250,12 @@ namespace Audit.EntityFramework
                     {
                         efEntry.ColumnValues[fk.Key] = fk.Value;
                     }
+
+                    var change = efEntry.Changes?.FirstOrDefault(e => e.ColumnName == fk.Key);
+                    if (change != null)
+                    {
+                        change.NewValue = fk.Value;
+                    }
                 }
             }
             // Update associations