+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to Audit.NET and its extensions will be documented in this f
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [19.1.1] - 2022-04-28
+- Audit.EntityFramework / Audit.EntityFramework.Core: `Changes.NewValue` property on the EF Audit Event was not being updated with the final value from the database 
+when the value changed was a foreign key to a new entity (#488)
+
 ## [19.1.0] - 2022-04-10
 - Audit.NET.JsonNewtonsoftAdapter: New `AuditContractResolver` configured by default on the Newtonsoft.Json adapter to enhance 
 compatibility when targeting newer frameworks (.NET >= 5.0) and still using Newtonsoft.Json as the serializer.